refactor(timeline): split timeline.tsx into a timeline/ folder
timeline.tsx had grown to 2,098 lines. Split per concern, following the layout
ui/core/wasm/bulk_sim/ already uses: an index.ts re-export keeps every import
site resolving through directory resolution, so detailed_results.tsx is
unchanged.

  timeline.tsx          1,197  the Timeline component
  constants.ts            334  colours, caches, action categories and id maps
  chart_series.ts         311  series, axes and annotations for the chart
  tooltip_content.tsx     154  tooltip bodies
  windowed_row.ts          64  horizontal windowing
  tooltips.ts              37  the per-row tippy delegate
  chart_view_picker.ts     27  the radio group behind the old select
  types.ts                 22
  index.ts                  1

The extracted chart and tooltip builders were private methods that only reached
for `this` to call each other, so they became free functions unchanged.
addResourceSeries is the exception: it read this.secondaryResource, which is a
parameter now.

Verified by comparing declaration-name sets before and after rather than by
compiling alone - 211 before, none missing - and in the browser: nothing built
while the tabs are hidden, 96 rotation rows with the ruler drawn, cast, aura,
tick and resource tooltips all resolving, no item mounted outside the window
after scrolling with zero DOM-order inversions, the chart's seven series with
threat hidden and the background traces dashed, the toolbar present, and the
chart tooltip rendering. No console errors.

Unused imports left behind by the move are removed.
