+++ ui/core/components/detailed_results/timeline/timeline.tsx
@@ -1,204 +1,45 @@
 import clsx from 'clsx';
-import tippy, { delegate, followCursor, Instance, Props } from 'tippy.js';
+import tippy from 'tippy.js';
 import { ref } from 'tsx-vanilla';
 
-import i18n from '../../../i18n/config';
-import { CacheHandler } from '../../cache_handler';
-import { APLActionItemSwap_SwapSet } from '../../proto/apl';
-import { OtherAction } from '../../proto/common';
-import { ResourceType } from '../../proto/spell';
-import { ActionId, buffAuraToSpellIdMap, resourceTypeToIcon } from '../../proto_utils/action_id';
-import { AuraUptimeLog, CastLog, DpsLog, ResourceChangedLogGroup, SimLog, ThreatLogGroup } from '../../proto_utils/logs_parser';
-import { resourceColors, resourceNames } from '../../proto_utils/names';
-import SecondaryResource from '../../proto_utils/secondary_resource';
-import { UnitMetrics } from '../../proto_utils/sim_result';
-import { orderedResourceTypes } from '../../proto_utils/utils';
-import { TypedEvent } from '../../typed_event';
-import { bucket, distinct, fragmentToString, maxIndex, stringComparator } from '../../utils';
-import { actionColors } from './color_settings';
-import { ResultComponent, ResultComponentConfig, SimResultData } from './result_component';
-
-type TooltipHandler = (dataPointIndex: number) => Element;
-
-const dpsColor = '#ed5653';
-const manaColor = '#2E93fA';
-const threatColor = '#b56d07';
-
-// Bounded: this holds detached DOM, and an unbounded CacheHandler never evicts.
-const cachedSpellCastIcon = new CacheHandler<HTMLAnchorElement>({ keysToKeep: 512 });
-
-// How far beyond the viewport rows stay populated, so a fast horizontal scroll does not
-// show gaps before the next frame lands.
-const ROW_WINDOW_PADDING_PX = 600;
-
-// The rotation is drawn at a fixed 100px per second, so a 300s fight is a ~30,000px wide row
-// against a viewport a few hundred px across. Items are built when their x-range first comes
-// on screen and detached (not discarded) when it leaves, so scrolling back is free.
-class WindowedRow {
-	private readonly items: Array<{ left: number; right: number; build: () => Element; elem?: Element }> = [];
-	private readonly mounted = new Set<number>();
-	// Reused across frames: mount() runs for every row on every animation frame.
-	private readonly wanted = new Set<number>();
-	// Running max of `right` over items[0..i], so the back-scan can stop as soon as no
-	// earlier item can still reach the window. Without it a single fight-long aura makes
-	// `widest` the whole row and the binary search degenerates into a full scan every frame.
-	private readonly maxRightUpTo: Array<number> = [];
-	private sorted = false;
-
-	constructor(private readonly rowElem: Element) {}
-
-	add(left: number, width: number, build: () => Element) {
-		this.items.push({ left, right: left + width, build });
-		this.sorted = false;
-	}
-
-	private sort() {
-		// A DoT's ticks can outlast the next cast, so items are not added in x order.
-		this.items.sort((a, b) => a.left - b.left);
-		this.maxRightUpTo.length = 0;
-		let max = -Infinity;
-		for (const item of this.items) {
-			max = Math.max(max, item.right);
-			this.maxRightUpTo.push(max);
-		}
-		this.sorted = true;
-	}
-
-	mount(windowLeft: number, windowRight: number) {
-		if (!this.sorted) this.sort();
-
-		// First item whose left edge is past the window; everything at or after it is out.
-		let lo = 0;
-		let hi = this.items.length;
-		while (lo < hi) {
-			const mid = (lo + hi) >> 1;
-			if (this.items[mid].left <= windowRight) lo = mid + 1;
-			else hi = mid;
-		}
-
-		const wanted = this.wanted;
-		wanted.clear();
-		for (let i = lo - 1; i >= 0 && this.maxRightUpTo[i] >= windowLeft; i--) {
-			if (this.items[i].right >= windowLeft) wanted.add(i);
-		}
-
-		this.mounted.forEach(index => {
-			if (!wanted.has(index)) {
-				this.items[index].elem!.remove();
-				this.mounted.delete(index);
-			}
-		});
-		// Mount in ascending order. The scan above runs backwards so it can stop early, but
-		// these are absolutely positioned siblings that abut each other, so DOM order decides
-		// which one wins an overlapping pixel - and appending right-to-left let a later block
-		// sit on top of its neighbour and answer for their hovers too.
-		const ascending = [...wanted].sort((a, b) => a - b);
-		for (const index of ascending) {
-			if (this.mounted.has(index)) continue;
-			const item = this.items[index];
-			item.elem ??= item.build();
-			this.rowElem.appendChild(item.elem);
-			this.mounted.add(index);
-		}
-	}
-}
-
-// Tooltips are delegated per row rather than instantiated per element, and their content is
-// built on first hover. Before, a rotation created one tippy instance - and its listeners -
-// for every cast, tick, aura window and resource block, several thousand of them, each
-// carrying a fully built subtree for a tooltip that in almost every case is never shown.
-const THREAT_SERIES_NAME = i18n.t('results_tab.details.timeline.tooltips.threat');
-
-// A radio group behaving like the <select> it replaced, so the rest of the component keeps
-// reading and writing a single `value`.
-class ChartViewPicker {
-	private readonly inputs: Array<HTMLInputElement>;
-	// Each choice is an input plus its label, and both have to be hidden together. Collected
-	// once here rather than re-queried on every plot update.
-	private readonly optionElems = new Map<string, Array<HTMLElement>>();
-
-	constructor(private readonly rootElem: HTMLElement) {
-		this.inputs = [...rootElem.querySelectorAll<HTMLInputElement>('input[type="radio"]')];
-		for (const input of this.inputs) {
-			this.optionElems.set(input.value, [...rootElem.querySelectorAll<HTMLElement>(`.${input.value}-option`)]);
-		}
-	}
-
-	get value(): string {
-		return this.inputs.find(input => input.checked)?.value ?? '';
-	}
-
-	set value(next: string) {
-		for (const input of this.inputs) input.checked = input.value === next;
-	}
-
-	setOptionVisible(option: string, visible: boolean) {
-		for (const elem of this.optionElems.get(option) ?? []) elem.classList.toggle('hide', !visible);
-	}
-
-	onChange(callback: () => void) {
-		this.rootElem.addEventListener('change', callback);
-	}
-}
-
-const TOOLTIP_TARGET_ATTR = 'data-timeline-tooltip';
-const tooltipBuilders = new WeakMap<Element, () => Element>();
-
-// Marks an element as a tooltip target for the delegate on its row.
-function addTooltip(reference: Element, buildContent: () => Element) {
-	reference.setAttribute(TOOLTIP_TARGET_ATTR, '');
-	tooltipBuilders.set(reference, buildContent);
-}
-
-function delegateTooltips(container: Element): Instance {
-	const built = new WeakSet<Element>();
-	return delegate(container, {
-		target: `[${TOOLTIP_TARGET_ATTR}]`,
-		placement: 'bottom',
-		content: '',
-		// Anchor to the pointer rather than to the element. An aura or resource bar can span
-		// the whole fight - tens of thousands of pixels - and anchoring to a box that wide put
-		// the tooltip out by the far edge of it, nowhere near what was being hovered.
-		// 'horizontal' tracks along the row, which is the axis these bars are long on, while
-		// the vertical placement stays tied to the row itself.
-		plugins: [followCursor],
-		followCursor: 'horizontal',
-		// Flip and slide along both axes so a tall tooltip near an edge stays on screen
-		// instead of hanging off it.
-		popperOptions: {
-			modifiers: [
-				{ name: 'flip', options: { padding: 8 } },
-				{ name: 'preventOverflow', options: { padding: 8, altAxis: true, tether: false } },
-			],
-		},
-		onShow(instance) {
-			const reference = instance.reference;
-			const buildContent = tooltipBuilders.get(reference);
-			if (!buildContent) return false;
-			if (!built.has(reference)) {
-				built.add(reference);
-				instance.setContent(buildContent());
-			}
-			return undefined;
-		},
-	} as Partial<Props> & { target: string });
-}
-
-interface TimelineConfig extends ResultComponentConfig {
-	secondaryResource?: SecondaryResource | null;
-}
-
-interface RotationSlot {
-	key: string;
-	labels: Array<Node>;
-	timeline: Array<Node>;
-	hiddenIdsNodes: Array<Node>;
-	emitter: TypedEvent<void>;
-	resetCallbacks: Array<() => void>;
-	plotOptions: any;
-	// Rows whose contents are populated from the horizontal scroll position.
-	windowedRows: Array<WindowedRow>;
-}
+import i18n from '../../../../i18n/config';
+import { ResourceType } from '../../../proto/spell';
+import { ActionId, buffAuraToSpellIdMap, resourceTypeToIcon } from '../../../proto_utils/action_id';
+import { AuraUptimeLog, CastLog, ResourceChangedLogGroup } from '../../../proto_utils/logs_parser';
+import { resourceNames } from '../../../proto_utils/names';
+import SecondaryResource from '../../../proto_utils/secondary_resource';
+import { UnitMetrics } from '../../../proto_utils/sim_result';
+import { orderedResourceTypes } from '../../../proto_utils/utils';
+import { TypedEvent } from '../../../typed_event';
+import { bucket, fragmentToString, stringComparator } from '../../../utils';
+import { ResultComponent, SimResultData } from '../result_component';
+import {
+	addDpsSeries,
+	addDpsYAxis,
+	addMajorCooldownAnnotations,
+	addManaSeries,
+	addResourceSeries,
+	addThreatSeries,
+	addThreatYAxis,
+	attachUnmappedSeriesToFirstAxis,
+} from './chart_series';
+import { ChartViewPicker } from './chart_view_picker';
+import {
+	auraAsResource,
+	cachedSpellCastIcon,
+	DEFAULT_ACTION_CATEGORY,
+	idsToGroupForRotation,
+	idToCategoryMap,
+	MELEE_ACTION_CATEGORY,
+	percentageResources,
+	ROW_WINDOW_PADDING_PX,
+	SPELL_ACTION_CATEGORY,
+	THREAT_SERIES_NAME,
+} from './constants';
+import { resourceTooltipElem } from './tooltip_content';
+import { addTooltip, delegateTooltips } from './tooltips';
+import { RotationSlot, TimelineConfig, TooltipHandler } from './types';
+import { WindowedRow } from './windowed_row';
 
 export class Timeline extends ResultComponent {
 	private readonly dpsResourcesPlotElem: HTMLElement;
@@ -494,16 +335,16 @@ export class Timeline extends ResultComponent {
 				return;
 			}
 
-			const dpsData = this.addDpsSeries(player, options, '');
-			this.addDpsYAxis(dpsData.maxDps, options);
+			const dpsData = addDpsSeries(player, options, '');
+			addDpsYAxis(dpsData.maxDps, options);
 			tooltipHandlers.push(dpsData.tooltipHandler);
-			tooltipHandlers.push(this.addManaSeries(player, options));
-			tooltipHandlers.push(this.addThreatSeries(player, options, ''));
-			tooltipHandlers.push(...this.addResourceSeries(player, options));
-			this.attachUnmappedSeriesToFirstAxis(options);
+			tooltipHandlers.push(addManaSeries(player, options));
+			tooltipHandlers.push(addThreatSeries(player, options, ''));
+			tooltipHandlers.push(...addResourceSeries(player, options, this.secondaryResource));
+			attachUnmappedSeriesToFirstAxis(options);
 			tooltipHandlers = tooltipHandlers.filter(handler => !!handler);
 
-			this.addMajorCooldownAnnotations(player, options);
+			addMajorCooldownAnnotations(player, options);
 		} else {
 			this.setRotationOptionVisible(false);
 
@@ -515,324 +356,24 @@ export class Timeline extends ResultComponent {
 			if (this.chartPicker.value == 'dps') {
 				let maxDps = 0;
 				players.forEach(player => {
-					const dpsData = this.addDpsSeries(player, options, `var(--bs-${player.classColor}`);
+					const dpsData = addDpsSeries(player, options, `var(--bs-${player.classColor}`);
 					maxDps = Math.max(maxDps, dpsData.maxDps);
 					tooltipHandlers.push(dpsData.tooltipHandler);
 				});
-				this.addDpsYAxis(maxDps, options);
+				addDpsYAxis(maxDps, options);
 			} else {
 				// threat
 				let maxThreat = 0;
 				players.forEach(player => {
-					tooltipHandlers.push(this.addThreatSeries(player, options, player.classColor));
+					tooltipHandlers.push(addThreatSeries(player, options, player.classColor));
 					maxThreat = Math.max(maxThreat, player.maxThreat);
 				});
-				this.addThreatYAxis(maxThreat, options);
+				addThreatYAxis(maxThreat, options);
 			}
 		}
 
 		if (this.liveSlot?.key === key) this.liveSlot.plotOptions = options;
 		this.applyChartOptions(options);
-	}
-
-	private addDpsYAxis(maxDps: number, options: any) {
-		const dpsAxisMax = Math.ceil(maxDps / 100) * 100;
-		options.yaxis.push({
-			color: dpsColor,
-			seriesName: 'DPS',
-			min: 0,
-			max: dpsAxisMax,
-			tickAmount: 10,
-			decimalsInFloat: 0,
-			title: {
-				text: 'DPS',
-				style: {
-					color: dpsColor,
-				},
-			},
-			axisBorder: {
-				show: true,
-				color: dpsColor,
-			},
-			axisTicks: {
-				color: dpsColor,
-			},
-			labels: {
-				minWidth: 30,
-				style: {
-					colors: [dpsColor],
-				},
-			},
-		});
-	}
-
-	private addThreatYAxis(maxThreat: number, options: any) {
-		const axisMax = Math.ceil(maxThreat / 10000) * 10000;
-		options.yaxis.push({
-			color: threatColor,
-			seriesName: i18n.t('results_tab.details.timeline.tooltips.threat'),
-			min: 0,
-			max: axisMax,
-			tickAmount: 10,
-			decimalsInFloat: 0,
-			title: {
-				text: i18n.t('results_tab.details.timeline.tooltips.threat'),
-				style: {
-					color: threatColor,
-				},
-			},
-			axisBorder: {
-				show: true,
-				color: threatColor,
-			},
-			axisTicks: {
-				color: threatColor,
-			},
-			labels: {
-				minWidth: 30,
-				style: {
-					colors: [threatColor],
-				},
-			},
-		});
-	}
-
-	// Returns a function for drawing the tooltip, or null if no series was added.
-	private addDpsSeries(unit: UnitMetrics, options: any, colorOverride: string): { maxDps: number; tooltipHandler: TooltipHandler } {
-		const dpsLogs = unit.dpsLogs.filter(log => log.timestamp >= 0);
-
-		options.colors.push(colorOverride || dpsColor);
-		options.series.push({
-			name: 'DPS',
-			type: 'line',
-			data: dpsLogs.map(log => {
-				return {
-					x: log.timestamp,
-					y: log.dps,
-				};
-			}),
-		});
-
-		return {
-			maxDps: dpsLogs[maxIndex(dpsLogs.map(l => l.dps))!]?.dps,
-			tooltipHandler: (dataPointIndex: number) => {
-				const log = dpsLogs[dataPointIndex];
-				return this.dpsTooltip(log, true, unit, colorOverride);
-			},
-		};
-	}
-
-	// ApexCharts maps y-axes to series by name. With a single axis every series shares it, but
-	// as soon as there is more than one, a series with no matching entry crashes its
-	// axis-matching pass. Threat has no axis of its own in the single-player view and rode on
-	// the DPS axis, so keep it there explicitly rather than giving it a scale of its own.
-	private attachUnmappedSeriesToFirstAxis(options: any) {
-		if (options.yaxis.length <= 1) return;
-
-		const mapped = new Set<string>();
-		for (const axis of options.yaxis) {
-			for (const name of Array.isArray(axis.seriesName) ? axis.seriesName : [axis.seriesName]) {
-				if (name) mapped.add(name);
-			}
-		}
-
-		const unmapped = options.series.map((series: any) => series.name).filter((name: string) => !mapped.has(name));
-		if (unmapped.length == 0) return;
-
-		const first = options.yaxis[0];
-		const existing = Array.isArray(first.seriesName) ? first.seriesName : [first.seriesName];
-		first.seriesName = [...existing, ...unmapped];
-	}
-
-	// Every other resource the unit actually used, drawn as a faint background trace behind
-	// the DPS line. Mana keeps its own labelled axis; these share a hidden 0-100% axis,
-	// because a rage bar and a rune count have nothing comparable about their raw numbers and
-	// one visible axis per resource would bury the chart.
-	private addResourceSeries(unit: UnitMetrics, options: any): Array<TooltipHandler> {
-		const handlers: Array<TooltipHandler> = [];
-		const traceNames = new Set<string>();
-
-		// Health is only informative where staying alive is the job; on a damage spec it is a
-		// flat line that just adds a trace to read past.
-		const showHealth = !!unit.spec?.isTankSpec || !!unit.spec?.isHealingSpec;
-
-		for (const resourceType of orderedResourceTypes) {
-			// Mana is already plotted against its own axis.
-			if (resourceType == ResourceType.ResourceTypeMana) continue;
-			if (resourceType == ResourceType.ResourceTypeHealth && !showHealth) continue;
-
-			const resourceLogs = unit.groupedResourceLogs[resourceType].filter(log => log.timestamp >= 0);
-			if (resourceLogs.length == 0) continue;
-
-			// Resources that start empty - chi, combo points, runic power - have no useful
-			// maxValue on their groups and a valueBefore of 0 at t=0, so scaling off either
-			// gave nonsense percentages. Take the largest value the resource is ever declared
-			// or observed to hold, over the whole fight.
-			let resourceMax = 0;
-			for (const log of resourceLogs) {
-				resourceMax = Math.max(resourceMax, log.maxValue || 0, log.valueBefore, log.valueAfter);
-			}
-			if (resourceMax <= 0) continue;
-			let name = resourceNames.get(resourceType)!;
-			if (resourceType == ResourceType.ResourceTypeGenericResource && this.secondaryResource?.name) {
-				name = this.secondaryResource.name;
-			}
-
-			traceNames.add(name);
-			options.colors.push(resourceColors.get(resourceType));
-			options.series.push({
-				name: name,
-				type: 'line',
-				data: resourceLogs.map(log => ({
-					x: log.timestamp,
-					// Percent of that resource's own maximum: the only scale they share.
-					y: Number(((log.valueAfter / resourceMax) * 100).toFixed(2)),
-				})),
-			});
-			options.yaxis.push({
-				seriesName: name,
-				show: false,
-				min: 0,
-				max: 100,
-			});
-
-			handlers.push((dataPointIndex: number) => {
-				const log = resourceLogs[dataPointIndex];
-				return this.resourceTooltipElem(log, resourceMax, false);
-			});
-		}
-
-		if (traceNames.size > 0) {
-			// Thin and dashed so they read as context behind the DPS line rather than as peers
-			// of it. Both are per-series arrays, so they have to cover every series in order.
-			options.stroke = {
-				curve: 'straight',
-				width: options.series.map((series: any) => (traceNames.has(series.name) ? 1 : 2)),
-				dashArray: options.series.map((series: any) => (traceNames.has(series.name) ? 4 : 0)),
-			};
-		}
-
-		return handlers;
-	}
-
-	// Returns a function for drawing the tooltip, or null if no series was added.
-	private addManaSeries(unit: UnitMetrics, options: any): TooltipHandler | null {
-		const manaLogs = unit.groupedResourceLogs[ResourceType.ResourceTypeMana].filter(log => log.timestamp >= 0);
-		if (manaLogs.length == 0) {
-			return null;
-		}
-		const maxMana = manaLogs[0].valueBefore;
-
-		options.colors.push(manaColor);
-		options.series.push({
-			name: 'Mana',
-			type: 'line',
-			data: manaLogs.map(log => {
-				return {
-					x: log.timestamp,
-					y: log.valueAfter,
-				};
-			}),
-		});
-		options.yaxis.push({
-			seriesName: 'Mana',
-			opposite: true, // Appear on right side
-			min: 0,
-			max: maxMana,
-			tickAmount: 10,
-			title: {
-				text: 'Mana',
-				style: {
-					color: manaColor,
-				},
-			},
-			axisBorder: {
-				show: true,
-				color: manaColor,
-			},
-			axisTicks: {
-				color: manaColor,
-			},
-			labels: {
-				minWidth: 30,
-				style: {
-					colors: [manaColor],
-				},
-				formatter: (val: string) => {
-					const v = parseFloat(val);
-					return `${v.toFixed(0)} (${((v / maxMana) * 100).toFixed(0)}%)`;
-				},
-			},
-		} as any);
-
-		return (dataPointIndex: number) => {
-			const log = manaLogs[dataPointIndex];
-			return this.resourceTooltip(log, maxMana, true);
-		};
-	}
-
-	// Returns a function for drawing the tooltip, or null if no series was added.
-	private addThreatSeries(unit: UnitMetrics, options: any, colorOverride: string): TooltipHandler | null {
-		options.colors.push(colorOverride || threatColor);
-		options.series.push({
-			name: THREAT_SERIES_NAME,
-			type: 'line',
-			data: unit.threatLogs
-				.filter(log => log.timestamp >= 0)
-				.map(log => {
-					return {
-						x: log.timestamp,
-						y: log.threatAfter,
-					};
-				}),
-		});
-
-		return (dataPointIndex: number) => {
-			const log = unit.threatLogs[dataPointIndex];
-			return this.threatTooltip(log, true, unit, colorOverride);
-		};
-	}
-
-	private addMajorCooldownAnnotations(unit: UnitMetrics, options: any) {
-		const mcdLogs = unit.majorCooldownLogs;
-		const mcdAuraLogs = unit.majorCooldownAuraUptimeLogs;
-
-		// Figure out how much to vertically offset cooldown icons, for cooldowns
-		// used very close to each other. This is so the icons don't overlap.
-		const MAX_ALLOWED_DIST = 10;
-		const cooldownIconOffsets = mcdLogs.map(
-			(mcdLog, mcdIdx) => mcdLogs.filter((cdLog, cdIdx) => cdIdx < mcdIdx && cdLog.timestamp > mcdLog.timestamp - MAX_ALLOWED_DIST).length,
-		);
-
-		const distinctMcdAuras = distinct(mcdAuraLogs, (a, b) => a.actionId!.equalsIgnoringTag(b.actionId!));
-		// Sort by name so auras keep their same colors even if timings change.
-		distinctMcdAuras.sort((a, b) => stringComparator(a.actionId!.name, b.actionId!.name));
-		const mcdAuraColors = mcdAuraLogs.map(
-			mcdAuraLog => actionColors[distinctMcdAuras.findIndex(dAura => dAura.actionId!.equalsIgnoringTag(mcdAuraLog.actionId!))],
-		);
-
-		options.annotations = {
-			position: 'back',
-			xaxis: mcdAuraLogs.map((log, i) => {
-				return {
-					x: log.gainedAt,
-					x2: log.fadedAt,
-					fillColor: mcdAuraColors[i],
-				};
-			}),
-			points: mcdLogs.map((log, i) => {
-				return {
-					x: log.timestamp,
-					y: 0,
-					image: {
-						path: log.actionId!.iconUrl,
-						width: 20,
-						height: 20,
-						offsetY: cooldownIconOffsets[i] * -25,
-					},
-				};
-			}),
-		};
 	}
 
 	private clearRotationChart() {
@@ -1227,7 +768,7 @@ export class Timeline extends ResultComponent {
 						resourceElem.textContent = Math.floor(resourceLogGroup.valueAfter).toFixed(0);
 					}
 				}
-				addTooltip(resourceElem, () => this.resourceTooltipElem(resourceLogGroup, startValue(resourceLogGroup), false));
+				addTooltip(resourceElem, () => resourceTooltipElem(resourceLogGroup, startValue(resourceLogGroup), false));
 				return resourceElem;
 			});
 		});
@@ -1495,155 +1036,6 @@ export class Timeline extends ResultComponent {
 		ctx.stroke();
 	}
 
-	private dpsTooltip(log: DpsLog, _includeAuras: boolean, player: UnitMetrics, colorOverride: string) {
-		const showPlayerLabel = colorOverride != '';
-		return (
-			<div className="timeline-tooltip dps">
-				<div className="timeline-tooltip-header">
-					{showPlayerLabel ? (
-						<>
-							<img className="timeline-tooltip-icon" src="${player.iconUrl}" />
-							<span className="" style="color: ${colorOverride}">
-								{player.label}
-							</span>
-							<span> - </span>
-						</>
-					) : null}
-					<span className="bold">{log.timestamp.toFixed(2)}s</span>
-				</div>
-				<div className="timeline-tooltip-body">
-					<ul className="timeline-dps-events">{log.damageLogs.map(damageLog => this.tooltipLogItem(damageLog, damageLog.result()))}</ul>
-					<div className="timeline-tooltip-body-row">
-						<span className="series-color">
-							{i18n.t('results_tab.details.timeline.tooltips.dps')}: {log.dps.toFixed(2)}
-						</span>
-					</div>
-				</div>
-				{this.tooltipAurasSection(log)}
-			</div>
-		);
-	}
-
-	private threatTooltip(log: ThreatLogGroup, includeAuras: boolean, player: UnitMetrics, colorOverride: string) {
-		const showPlayerLabel = colorOverride != '';
-		return (
-			<div className="timeline-tooltip threat">
-				<div className="timeline-tooltip-header">
-					{showPlayerLabel ? (
-						<>
-							<img className="timeline-tooltip-icon" src={player.iconUrl} />
-							<span className="" style={{ color: colorOverride }}>
-								{player.label}
-							</span>
-							<span> - </span>
-						</>
-					) : null}
-					<span className="bold">{log.timestamp.toFixed(2)}s</span>
-				</div>
-				<div className="timeline-tooltip-body">
-					<div className="timeline-tooltip-body-row">
-						<span className="series-color">
-							{i18n.t('results_tab.details.timeline.tooltips.before')}: {log.threatBefore.toFixed(1)}
-						</span>
-					</div>
-					<ul className="timeline-threat-events">
-						{log.logs.map(log =>
-							this.tooltipLogItem(
-								log,
-								<>
-									{log.threat.toFixed(1)} {i18n.t('results_tab.details.timeline.tooltips.threat')}
-								</>,
-							),
-						)}
-					</ul>
-					<div className="timeline-tooltip-body-row">
-						<span className="series-color">
-							{i18n.t('results_tab.details.timeline.tooltips.after')}: {log.threatAfter.toFixed(1)}
-						</span>
-					</div>
-				</div>
-				{includeAuras ? this.tooltipAurasSection(log) : null}
-			</div>
-		);
-	}
-
-	private resourceTooltipElem(log: ResourceChangedLogGroup, maxValue: number, includeAuras: boolean) {
-		const valToDisplayString = percentageResources.includes(log.resourceType)
-			? (val: number) => `${val.toFixed(1)} (${((val / maxValue) * 100).toFixed(0)}%)`
-			: (val: number) => `${val.toFixed(1)}`;
-
-		return (
-			<div className={`timeline-tooltip ${resourceNames.get(log.resourceType)!.toLowerCase().replaceAll(' ', '-')}`}>
-				<div className="timeline-tooltip-header">
-					<span className="bold">{log.timestamp.toFixed(2)}s</span>
-				</div>
-				<div className="timeline-tooltip-body">
-					<div className="timeline-tooltip-body-row">
-						<span className="series-color">
-							{i18n.t('results_tab.details.timeline.tooltips.before')}: {valToDisplayString(log.valueBefore)}
-						</span>
-					</div>
-					<ul className="timeline-mana-events">
-						{log.logs.map(manaChangedLog => this.tooltipLogItemElem(manaChangedLog, <>{manaChangedLog.resultString()}</>))}
-					</ul>
-					<div className="timeline-tooltip-body-row">
-						<span className="series-color">
-							{i18n.t('results_tab.details.timeline.tooltips.after')}: {valToDisplayString(log.valueAfter)}
-						</span>
-					</div>
-				</div>
-				{includeAuras && this.tooltipAurasSectionElem(log)}
-			</div>
-		);
-	}
-
-	private resourceTooltip(log: ResourceChangedLogGroup, maxValue: number, includeAuras: boolean) {
-		return this.resourceTooltipElem(log, maxValue, includeAuras);
-	}
-
-	private tooltipLogItem(log: SimLog, value: Element) {
-		return this.tooltipLogItemElem(log, value);
-	}
-
-	private tooltipLogItemElem(log: SimLog, value: Element): JSX.Element {
-		return (
-			<li>
-				{log.actionId && log.actionId.iconUrl && <img className="timeline-tooltip-icon" src={log.actionId.iconUrl}></img>}
-				{log.actionId && <span>{log.actionId.name}</span>}
-				<span className="series-color">{value}</span>
-			</li>
-		);
-	}
-
-	private tooltipAurasSection(log: SimLog) {
-		if (log.activeAuras.length == 0) {
-			return '';
-		}
-		return this.tooltipAurasSectionElem(log);
-	}
-
-	private tooltipAurasSectionElem(log: SimLog): JSX.Element {
-		if (log.activeAuras.length == 0) {
-			return <></>;
-		}
-
-		return (
-			<div className="timeline-tooltip-auras">
-				<div className="timeline-tooltip-body-row">
-					<span className="bold">{i18n.t('results_tab.details.timeline.tooltips.active_auras')}</span>
-				</div>
-				<ul className="timeline-active-auras">
-					{log.activeAuras.map(auraLog => (
-						<li>
-							{auraLog.actionId!.iconUrl && <img className="timeline-tooltip-icon" src={auraLog.actionId!.iconUrl}></img>}
-							<span>{auraLog.actionId!.name}</span>
-						</li>
-					))}
-				</ul>
-			</div>
-		);
-	}
-
 	update() {
 		this.updatePlot();
 	}
@@ -1803,321 +1195,3 @@ export class Timeline extends ResultComponent {
 		super.reset();
 	}
 }
-
-const MELEE_ACTION_CATEGORY = 1;
-const SPELL_ACTION_CATEGORY = 2;
-const DEFAULT_ACTION_CATEGORY = 3;
-
-const auraAsResource: ActionId[] = [
-	// APL Damage Amplifier
-	ActionId.fromOtherId(OtherAction.OtherActionDamageAmplifier),
-
-	// Item Swap
-	ActionId.fromOtherId(OtherAction.OtherActionItemSwap, APLActionItemSwap_SwapSet.Main),
-	ActionId.fromOtherId(OtherAction.OtherActionItemSwap, APLActionItemSwap_SwapSet.Swap1),
-
-	// Vengeance
-	ActionId.fromSpellId(84840), // Druid
-	ActionId.fromSpellId(84839), // Paladin
-	ActionId.fromSpellId(93098), // Warrior
-	ActionId.fromSpellId(93099), // Death Knight
-	ActionId.fromSpellId(120267), // Monk
-
-	// Monk
-	ActionId.fromSpellId(124255, 1), // Stagger
-	ActionId.fromSpellId(128938), // Elusive Brew - Stacks
-	ActionId.fromSpellId(115308), // Elusive Brew - Active
-	ActionId.fromSpellId(1247279), // Tiger Eye Brew - Stacks
-	ActionId.fromSpellId(1247275), // Tiger Eye Brew - Active
-
-	// Mage
-	ActionId.fromSpellId(148022), // Icicle
-];
-
-// Hard-coded spell categories for controlling rotation ordering.
-const idToCategoryMap: Record<number, number> = {
-	[OtherAction.OtherActionMove]: 0,
-	[OtherAction.OtherActionAttack]: 0.01,
-	[OtherAction.OtherActionShoot]: 0.5,
-	[OtherAction.OtherActionDamageAmplifier]: 0.6,
-
-	// Druid
-	[48480]: 0.1, // Maul
-	[48564]: MELEE_ACTION_CATEGORY + 0.1, // Mangle (Bear)
-	[48568]: MELEE_ACTION_CATEGORY + 0.2, // Lacerate
-	[48562]: MELEE_ACTION_CATEGORY + 0.3, // Swipe (Bear)
-
-	[48566]: MELEE_ACTION_CATEGORY + 0.1, // Mangle (Cat)
-	[48572]: MELEE_ACTION_CATEGORY + 0.2, // Shred
-	[49800]: MELEE_ACTION_CATEGORY + 0.51, // Rip
-	[52610]: MELEE_ACTION_CATEGORY + 0.52, // Savage Roar
-	[48577]: MELEE_ACTION_CATEGORY + 0.53, // Ferocious Bite
-
-	[48465]: SPELL_ACTION_CATEGORY + 0.1, // Starfire
-	[48461]: SPELL_ACTION_CATEGORY + 0.2, // Wrath
-	[53201]: SPELL_ACTION_CATEGORY + 0.3, // Starfall
-	[48463]: SPELL_ACTION_CATEGORY + 0.4, // Moonfire
-
-	// Hunter
-	[48996]: 0.1, // Raptor Strike
-	[53217]: 0.6, // Wild Quiver
-	[53209]: MELEE_ACTION_CATEGORY + 0.1, // Chimera Shot
-	[53353]: MELEE_ACTION_CATEGORY + 0.11, // Chimera Shot Serpent
-	[53301]: MELEE_ACTION_CATEGORY + 0.1, // Explosive Shot
-	[1215485]: MELEE_ACTION_CATEGORY + 0.12, // Explosive Shot
-	[49050]: MELEE_ACTION_CATEGORY + 0.2, // Aimed Shot
-	[49048]: MELEE_ACTION_CATEGORY + 0.21, // Multi Shot
-	[3044]: MELEE_ACTION_CATEGORY + 0.22, // Arcane Shot
-	[56641]: MELEE_ACTION_CATEGORY + 0.27, // Steady Shot
-	[53351]: MELEE_ACTION_CATEGORY + 0.28, // Kill Shot
-	[34490]: MELEE_ACTION_CATEGORY + 0.29, // Silencing Shot
-	[49001]: MELEE_ACTION_CATEGORY + 0.3, // Serpent Sting
-	[53238]: MELEE_ACTION_CATEGORY + 0.31, // Piercing Shots
-	[3674]: MELEE_ACTION_CATEGORY + 0.32, // Black Arrow
-	[49067]: MELEE_ACTION_CATEGORY + 0.33, // Explosive Trap
-	[77767]: MELEE_ACTION_CATEGORY + 0.34, // Cobra Shot
-
-	// Paladin
-	[76672]: MELEE_ACTION_CATEGORY + 0.01, // Hand of Light (mastery)
-	[35395]: MELEE_ACTION_CATEGORY + 0.02, // Crusader Strike
-	[53595]: MELEE_ACTION_CATEGORY + 0.04, // Hammer of the Righteous (Physical)
-	[88263]: MELEE_ACTION_CATEGORY + 0.05, // Hammer of the Righteous (Holy)
-	[53385]: MELEE_ACTION_CATEGORY + 0.06, // Divine Storm
-	[85256]: MELEE_ACTION_CATEGORY + 0.07, // Templar's Verdict
-	[20271]: MELEE_ACTION_CATEGORY + 0.08, // Judgment
-	[42463]: MELEE_ACTION_CATEGORY + 0.09, // Seal of Truth (on-hit)
-	[31803]: MELEE_ACTION_CATEGORY + 0.1, // Censure (Seal of Truth)
-	[101423]: MELEE_ACTION_CATEGORY + 0.11, // Seal of Righteousness
-	[53600]: MELEE_ACTION_CATEGORY + 0.12, // Shield of the Righteous
-	[879]: MELEE_ACTION_CATEGORY + 0.15, // Exorcism
-	[26573]: MELEE_ACTION_CATEGORY + 0.16, // Consecration
-	[119072]: MELEE_ACTION_CATEGORY + 0.17, // Holy Wrath
-	[24275]: MELEE_ACTION_CATEGORY + 0.18, // Hammer of Wrath
-	[114852]: MELEE_ACTION_CATEGORY + 0.19, // Holy Prism (Damage)
-	[114919]: MELEE_ACTION_CATEGORY + 0.19, // Arcing Light (Damage)
-	[114916]: MELEE_ACTION_CATEGORY + 0.19, // Execution Sentence
-	[114871]: MELEE_ACTION_CATEGORY + 0.2, // Holy Prism (Heal)
-	[119952]: MELEE_ACTION_CATEGORY + 0.2, // Arcing Light (Heal)
-	[146586]: MELEE_ACTION_CATEGORY + 0.2, // Stay of Execution
-	[84963]: SPELL_ACTION_CATEGORY + 0.01, // Inquisition
-	[54428]: SPELL_ACTION_CATEGORY + 0.02, // Divine Plea
-	[498]: SPELL_ACTION_CATEGORY + 0.03, // Divine Protection
-	[66233]: SPELL_ACTION_CATEGORY + 0.05, // Ardent Defender
-	[31884]: SPELL_ACTION_CATEGORY + 0.06, // Avenging Wrath
-	[114232]: SPELL_ACTION_CATEGORY + 0.07, // Sanctified Wrath
-	[105809]: SPELL_ACTION_CATEGORY + 0.08, // Holy Avenger,
-	[86698]: SPELL_ACTION_CATEGORY + 0.09, // Guardian of Ancient Kings
-	[86704]: SPELL_ACTION_CATEGORY + 0.1, // Ancient Fury
-	[20925]: SPELL_ACTION_CATEGORY + 0.11, // Sacred Shield (Ret / Prot)
-	[148039]: SPELL_ACTION_CATEGORY + 0.11, // Sacred Shield (Holy)
-	[65148]: SPELL_ACTION_CATEGORY + 0.12, // Sacred Shield (Absorb)
-	[114039]: SPELL_ACTION_CATEGORY + 0.13, // Hand of Purity
-
-	// Priest
-	[48300]: SPELL_ACTION_CATEGORY + 0.11, // Devouring Plague
-	[48125]: SPELL_ACTION_CATEGORY + 0.12, // Shadow Word: Pain
-	[48160]: SPELL_ACTION_CATEGORY + 0.13, // Vampiric Touch
-	[48135]: SPELL_ACTION_CATEGORY + 0.14, // Holy Fire
-	[48123]: SPELL_ACTION_CATEGORY + 0.19, // Smite
-	[48127]: SPELL_ACTION_CATEGORY + 0.2, // Mind Blast
-	[48158]: SPELL_ACTION_CATEGORY + 0.3, // Shadow Word: Death
-	[48156]: SPELL_ACTION_CATEGORY + 0.4, // Mind Flay
-
-	// Rogue
-	[6774]: MELEE_ACTION_CATEGORY + 0.1, // Slice and Dice
-	[8647]: MELEE_ACTION_CATEGORY + 0.2, // Expose Armor
-	[48672]: MELEE_ACTION_CATEGORY + 0.3, // Rupture
-	[57993]: MELEE_ACTION_CATEGORY + 0.3, // Envenom
-	[48668]: MELEE_ACTION_CATEGORY + 0.4, // Eviscerate
-	[48666]: MELEE_ACTION_CATEGORY + 0.5, // Mutilate
-	[48665]: MELEE_ACTION_CATEGORY + 0.6, // Mutilate (MH)
-	[48664]: MELEE_ACTION_CATEGORY + 0.7, // Mutilate (OH)
-	[48638]: MELEE_ACTION_CATEGORY + 0.5, // Sinister Strike
-	[51723]: MELEE_ACTION_CATEGORY + 0.8, // Fan of Knives
-	[57973]: SPELL_ACTION_CATEGORY + 0.1, // Deadly Poison
-	[57968]: SPELL_ACTION_CATEGORY + 0.2, // Instant Poison
-
-	// Shaman
-	[8232]: 0.11, // Windfury Weapon
-	[8024]: 0.12, // Flametongue Weapon
-	[8033]: 0.12, // Frostbrand Weapon
-	[17364]: MELEE_ACTION_CATEGORY + 0.1, // Stormstrike
-	[60103]: MELEE_ACTION_CATEGORY + 0.2, // Lava Lash
-	[49233]: SPELL_ACTION_CATEGORY + 0.21, // Flame Shock
-	[49231]: SPELL_ACTION_CATEGORY + 0.22, // Earth Shock
-	[49236]: SPELL_ACTION_CATEGORY + 0.23, // Frost Shock
-	[60043]: SPELL_ACTION_CATEGORY + 0.31, // Lava Burst
-	[49238]: SPELL_ACTION_CATEGORY + 0.32, // Lightning Bolt
-	[49271]: SPELL_ACTION_CATEGORY + 0.33, // Chain Lightning
-	[61657]: SPELL_ACTION_CATEGORY + 0.41, // Fire Nova
-	[58734]: SPELL_ACTION_CATEGORY + 0.42, // Magma Totem
-	[58704]: SPELL_ACTION_CATEGORY + 0.43, // Searing Totem
-	[49281]: SPELL_ACTION_CATEGORY + 0.51, // Lightning Shield
-	[49279]: SPELL_ACTION_CATEGORY + 0.52, // Lightning Shield (Proc)
-	[2825]: DEFAULT_ACTION_CATEGORY + 0.1, // Bloodlust
-
-	// Warlock
-	[603]: SPELL_ACTION_CATEGORY + 0.01, // Curse of Doom
-	[980]: SPELL_ACTION_CATEGORY + 0.02, // Curse of Agony
-	[172]: SPELL_ACTION_CATEGORY + 0.1, // Corruption
-	[48181]: SPELL_ACTION_CATEGORY + 0.2, // Haunt
-	[30108]: SPELL_ACTION_CATEGORY + 0.3, // Unstable Affliction
-	[348]: SPELL_ACTION_CATEGORY + 0.31, // Immolate
-	[17962]: SPELL_ACTION_CATEGORY + 0.32, // Conflagrate
-	[50796]: SPELL_ACTION_CATEGORY + 0.49, // Chaos Bolt
-	[686]: SPELL_ACTION_CATEGORY + 0.5, // Shadow Bolt
-	[29722]: SPELL_ACTION_CATEGORY + 0.51, // Incinerate
-	[6353]: SPELL_ACTION_CATEGORY + 0.52, // Soul Fire
-	[1120]: SPELL_ACTION_CATEGORY + 0.6, // Drain Soul
-	[1454]: SPELL_ACTION_CATEGORY + 0.7, // Life Tap
-	[59672]: SPELL_ACTION_CATEGORY + 0.8, // Metamorphosis
-	[104025]: SPELL_ACTION_CATEGORY + 0.81, // Immolation Aura
-	[129476]: SPELL_ACTION_CATEGORY + 0.81, // Immolation Aura
-	[47193]: SPELL_ACTION_CATEGORY + 0.82, // Demonic Empowerment
-
-	// Mage
-	[42842]: SPELL_ACTION_CATEGORY + 0.01, // Frostbolt
-	[47610]: SPELL_ACTION_CATEGORY + 0.02, // Frostfire Bolt
-	[42897]: SPELL_ACTION_CATEGORY + 0.02, // Arcane Blast
-	[42833]: SPELL_ACTION_CATEGORY + 0.02, // Fireball
-	[10]: SPELL_ACTION_CATEGORY + 0.021, // Blizzard - Cast
-	[42208]: SPELL_ACTION_CATEGORY + 0.022, // Blizzard - Tick
-	[42859]: SPELL_ACTION_CATEGORY + 0.03, // Scorch
-	[42891]: SPELL_ACTION_CATEGORY + 0.1, // Pyroblast
-	[42846]: SPELL_ACTION_CATEGORY + 0.1, // Arcane Missiles
-	[44572]: SPELL_ACTION_CATEGORY + 0.1, // Deep Freeze
-	[44781]: SPELL_ACTION_CATEGORY + 0.2, // Arcane Barrage
-	[42914]: SPELL_ACTION_CATEGORY + 0.2, // Ice Lance
-	[55360]: SPELL_ACTION_CATEGORY + 0.2, // Living Bomb
-	[55362]: SPELL_ACTION_CATEGORY + 0.21, // Living Bomb (Explosion)
-	[12654]: SPELL_ACTION_CATEGORY + 0.3, // Ignite
-	[12472]: SPELL_ACTION_CATEGORY + 0.4, // Icy Veins
-	[11129]: SPELL_ACTION_CATEGORY + 0.4, // Combustion
-	[12042]: SPELL_ACTION_CATEGORY + 0.4, // Arcane Power
-	[11958]: SPELL_ACTION_CATEGORY + 0.41, // Cold Snap
-	[12043]: SPELL_ACTION_CATEGORY + 0.41, // Presence of Mind
-	[31687]: SPELL_ACTION_CATEGORY + 0.41, // Water Elemental
-	[55342]: SPELL_ACTION_CATEGORY + 0.5, // Mirror Image
-	[33312]: SPELL_ACTION_CATEGORY + 0.51, // Mana Gems
-	[12051]: SPELL_ACTION_CATEGORY + 0.52, // Evocate
-	[44401]: SPELL_ACTION_CATEGORY + 0.6, // Missile Barrage
-	[44448]: SPELL_ACTION_CATEGORY + 0.6, // Hot Streak
-	[44545]: SPELL_ACTION_CATEGORY + 0.6, // Fingers of Frost
-	[44549]: SPELL_ACTION_CATEGORY + 0.61, // Brain Freeze
-	[12536]: SPELL_ACTION_CATEGORY + 0.61, // Clearcasting
-
-	// Warrior
-	[47520]: 0.1, // Cleave
-	[47450]: 0.1, // Heroic Strike
-	[47475]: MELEE_ACTION_CATEGORY + 0.05, // Slam
-	[23881]: MELEE_ACTION_CATEGORY + 0.1, // Bloodthirst
-	[47486]: MELEE_ACTION_CATEGORY + 0.1, // Mortal Strike
-	[30356]: MELEE_ACTION_CATEGORY + 0.1, // Shield Slam
-	[47498]: MELEE_ACTION_CATEGORY + 0.21, // Devastate
-	[47467]: MELEE_ACTION_CATEGORY + 0.22, // Sunder Armor
-	[57823]: MELEE_ACTION_CATEGORY + 0.23, // Revenge
-	[1680]: MELEE_ACTION_CATEGORY + 0.24, // Whirlwind
-	[7384]: MELEE_ACTION_CATEGORY + 0.25, // Overpower
-	[47471]: MELEE_ACTION_CATEGORY + 0.42, // Execute
-	[12867]: SPELL_ACTION_CATEGORY + 0.51, // Deep Wounds
-	[58874]: SPELL_ACTION_CATEGORY + 0.52, // Damage Shield
-	[47296]: SPELL_ACTION_CATEGORY + 0.53, // Critical Block
-	[46924]: MELEE_ACTION_CATEGORY + 0.61, // Bladestorm
-	[46968]: MELEE_ACTION_CATEGORY + 0.61, // Shockwave
-	[118000]: MELEE_ACTION_CATEGORY + 0.61, // Dragon Roar
-	[2565]: SPELL_ACTION_CATEGORY + 0.62, // Shield Block
-	[112048]: SPELL_ACTION_CATEGORY + 0.63, // Shield Barrier
-	[76857]: SPELL_ACTION_CATEGORY + 0.64, // Mastery: Critical Block
-	[1249459]: SPELL_ACTION_CATEGORY + 0.65, // Shattering Throw
-	[71]: DEFAULT_ACTION_CATEGORY + 0.1, // Defensive Stance
-	[2457]: DEFAULT_ACTION_CATEGORY + 0.1, // Battle Stance
-	[6673]: DEFAULT_ACTION_CATEGORY + 0.1, // Battle Shout
-	[469]: DEFAULT_ACTION_CATEGORY + 0.1, // Commanding Shout
-
-	// Death Knight
-	[49998]: MELEE_ACTION_CATEGORY + 0.01, // Death Strike
-	[45470]: MELEE_ACTION_CATEGORY + 0.02, // Death Strike (Heal)
-	[77535]: MELEE_ACTION_CATEGORY + 0.03, // Blood Shield
-	[49184]: MELEE_ACTION_CATEGORY + 0.04, // Howling Blast
-	[49020]: MELEE_ACTION_CATEGORY + 0.05, // Obliterate
-	[49143]: MELEE_ACTION_CATEGORY + 0.1, // Frost strike
-	[45902]: MELEE_ACTION_CATEGORY + 0.15, // Blood strike
-	[50842]: MELEE_ACTION_CATEGORY + 0.2, // Pestilence
-	[47541]: MELEE_ACTION_CATEGORY + 0.25, // Death Coil
-	[43265]: MELEE_ACTION_CATEGORY + 0.25, // Death and Decay
-	[63560]: MELEE_ACTION_CATEGORY + 0.25, // Dark Transformation
-	[50536]: MELEE_ACTION_CATEGORY + 0.25, // Unholy Blight
-	[57623]: MELEE_ACTION_CATEGORY + 0.25, // HoW
-	[45477]: MELEE_ACTION_CATEGORY + 0.3, // Icy touch
-	[45462]: MELEE_ACTION_CATEGORY + 0.3, // Plague strike
-	[114866]: MELEE_ACTION_CATEGORY + 0.31, // Soul Reaper
-	[130735]: MELEE_ACTION_CATEGORY + 0.31, // Soul Reaper
-	[130736]: MELEE_ACTION_CATEGORY + 0.31, // Soul Reaper
-	[114867]: MELEE_ACTION_CATEGORY + 0.32, // Soul Reaper (Tick)
-	[51271]: MELEE_ACTION_CATEGORY + 0.35, // UA
-	[45529]: MELEE_ACTION_CATEGORY + 0.35, // BT
-	[47568]: MELEE_ACTION_CATEGORY + 0.35, // ERW
-	[49206]: MELEE_ACTION_CATEGORY + 0.35, // Summon Gargoyle
-	[46584]: MELEE_ACTION_CATEGORY + 0.35, // Raise Dead
-	[55095]: MELEE_ACTION_CATEGORY + 0.4, // Frost Fever
-	[55078]: MELEE_ACTION_CATEGORY + 0.4, // Blood Plague
-	[50401]: MELEE_ACTION_CATEGORY + 0.5, // Razor Frost
-	[50689]: DEFAULT_ACTION_CATEGORY + 0.1, // Blood Presence
-	[48263]: DEFAULT_ACTION_CATEGORY + 0.1, // Frost Presence
-	[48265]: DEFAULT_ACTION_CATEGORY + 0.1, // Unholy Presence
-
-	// Monk
-	[120274]: 0.02, // Tiger Strikes (Main Hand)
-	[120278]: 0.03, // Tiger Strikes (Off Hand)
-	[100780]: MELEE_ACTION_CATEGORY + 0.01, // Jab
-	[100787]: MELEE_ACTION_CATEGORY + 0.02, // Tiger Palm
-	[100784]: MELEE_ACTION_CATEGORY + 0.03, // Blackout Kick
-	[130320]: MELEE_ACTION_CATEGORY + 0.04, // Rising Sun Kick
-	[113656]: MELEE_ACTION_CATEGORY + 0.05, // Fists of Fury (Cast)
-	[117418]: MELEE_ACTION_CATEGORY + 0.06, // Fists of Fury (Hit)
-	[101546]: MELEE_ACTION_CATEGORY + 0.07, // Spinning Crane Kick (Cast)
-	[107270]: MELEE_ACTION_CATEGORY + 0.08, // Spinning Crane Kick (Hit)
-	[116847]: MELEE_ACTION_CATEGORY + 0.07, // Rushing Jade Wind (Cast)
-	[148187]: MELEE_ACTION_CATEGORY + 0.08, // Rushing Jade Wind (Hit)
-	[115098]: SPELL_ACTION_CATEGORY + 0.01, // Chi Wave
-	[132467]: SPELL_ACTION_CATEGORY + 0.011, // Chi Wave (Damage)
-	[132463]: SPELL_ACTION_CATEGORY + 0.012, // Chi Wave (Heal)
-	[124098]: SPELL_ACTION_CATEGORY + 0.01, // Zen Sphere (Damage)
-	[124081]: SPELL_ACTION_CATEGORY + 0.011, // Zen Sphere (Heal)
-	[125033]: SPELL_ACTION_CATEGORY + 0.011, // Zen Sphere: Detonate (Damage)
-	[124101]: SPELL_ACTION_CATEGORY + 0.011, // Zen Sphere: Detonate (Heal)
-	[123986]: SPELL_ACTION_CATEGORY + 0.01, // Chi Burst
-	[148135]: SPELL_ACTION_CATEGORY + 0.011, // Chi Burst (Damage)
-	[130654]: SPELL_ACTION_CATEGORY + 0.012, // Chi Burst (Heal)
-	[1247275]: SPELL_ACTION_CATEGORY + 0.02, // Tigereye Brew
-	[115399]: SPELL_ACTION_CATEGORY + 0.03, // Chi Brew
-	[115288]: SPELL_ACTION_CATEGORY + 0.04, // Energizing Brew
-	[123402]: SPELL_ACTION_CATEGORY + 0.04, // Guard
-	[115295]: SPELL_ACTION_CATEGORY + 0.04, // Guard
-	[126456]: SPELL_ACTION_CATEGORY + 0.05, // Fortifying Brew
-	[123904]: SPELL_ACTION_CATEGORY + 0.06, // Invoke Xuen, the White Tiger
-	[115008]: SPELL_ACTION_CATEGORY + 0.06, // Chi Torpedo
-
-	// Generic
-	[53307]: SPELL_ACTION_CATEGORY + 0.931, // Thorns
-	[54043]: SPELL_ACTION_CATEGORY + 0.932, // Retribution Aura
-	[54758]: SPELL_ACTION_CATEGORY + 0.933, // Hyperspeed Acceleration
-	[42641]: SPELL_ACTION_CATEGORY + 0.941, // Sapper
-	[40536]: SPELL_ACTION_CATEGORY + 0.942, // Explosive Decoy
-	[41119]: SPELL_ACTION_CATEGORY + 0.943, // Saronite Bomb
-	[40771]: SPELL_ACTION_CATEGORY + 0.944, // Cobalt Frag Bomb
-	[120687]: DEFAULT_ACTION_CATEGORY + 0.945, // Stormlash Totem
-	[114206]: DEFAULT_ACTION_CATEGORY + 0.946, // Skull Bnaner
-};
-
-const idsToGroupForRotation: Array<number> = [
-	5171, // Rogue - Slice and Dice
-	2098, // Rogue - Eviscerate
-	1943, // Rogue - Rupture
-	51690, // Rogue - Killing Spree
-	32645, // Rogue - Envenom
-	16511, // Rogue - Hemorrhage
-	121471, // Rogue - Shadow Blades
-];
-
-const percentageResources: Array<ResourceType> = [ResourceType.ResourceTypeHealth, ResourceType.ResourceTypeMana];
